chore(example): update example file structure (#77)
* chore(example): update example file structure to make it compatible with the tester

* fix(example): lint and type errors
